Module google_reseller1_sandbox::api

Structs

Address
JSON template for address of a customer.

ChangePlanRequest
JSON template for the ChangePlan rpc request.

Customer
JSON template for a customer.

CustomerGetCall
Gets a customer resource if one exists and is owned by the reseller.

CustomerInsertCall
Creates a customer resource if one does not already exist.

CustomerMethods
A builder providing access to all methods supported on customer resources. It is not used directly, but through the Reseller hub.

CustomerPatchCall
Update a customer resource if one it exists and is owned by the reseller. This method supports patch semantics.

CustomerUpdateCall
Update a customer resource if one it exists and is owned by the reseller.

RenewalSettings
JSON template for a subscription renewal settings.

Reseller
Central instance to access all Reseller related resource activities

Seats
JSON template for subscription seats.

Subscription
JSON template for a subscription.

SubscriptionActivateCall
Activates a subscription previously suspended by the reseller

SubscriptionChangePlanCall
Changes the plan of a subscription

SubscriptionChangeRenewalSettingCall
Changes the renewal settings of a subscription

SubscriptionChangeSeatCall
Changes the seats configuration of a subscription

SubscriptionDeleteCall
Cancels/Downgrades a subscription.

SubscriptionGetCall
Gets a subscription of the customer.

SubscriptionInsertCall
Creates/Transfers a subscription for the customer.

SubscriptionListCall
Lists subscriptions of a reseller, optionally filtered by a customer name prefix.

SubscriptionMethods
A builder providing access to all methods supported on subscription resources. It is not used directly, but through the Reseller hub.

SubscriptionPlan
Plan details of the subscription

SubscriptionPlanCommitmentInterval
Interval of the commitment if it is a commitment plan.

SubscriptionStartPaidServiceCall
Starts paid service of a trial subscription

SubscriptionSuspendCall
Suspends an active subscription

SubscriptionTransferInfo
Transfer related information for the subscription.

SubscriptionTrialSettings
Trial Settings of the subscription.

Subscriptions
JSON template for a subscription list.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.