Struct google_reseller1_sandbox::ChangePlanRequest

pub struct ChangePlanRequest {
    pub plan_name: Option<String>,
    pub kind: Option<String>,
    pub purchase_order_id: Option<String>,
    pub seats: Option<Seats>,
}

JSON template for the ChangePlan rpc request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

change plan subscriptions (request)

Fields

plan_name

Name of the plan to change to.

kind

Identifies the resource as a subscription change plan request.

purchase_order_id

Purchase order id for your order tracking purposes.

seats

Number/Limit of seats in the new plan.