Struct google_reseller1_sandbox::Customer

pub struct Customer {
    pub customer_domain: Option<String>,
    pub customer_domain_verified: Option<bool>,
    pub alternate_email: Option<String>,
    pub kind: Option<String>,
    pub resource_ui_url: Option<String>,
    pub phone_number: Option<String>,
    pub postal_address: Option<Address>,
    pub customer_id: Option<String>,
}

JSON template for a customer.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update customers (request|response)

patch customers (request|response)

insert customers (request|response)

get customers (response)

Fields
customer_domain: Option<String>

The domain name of the customer.

customer_domain_verified: Option<bool>

Whether the customer's primary domain has been verified.

alternate_email: Option<String>

The alternate email of the customer.

kind: Option<String>

Identifies the resource as a customer.

resource_ui_url: Option<String>

Ui url for customer resource.

phone_number: Option<String>

The phone number of the customer.

postal_address: Option<Address>

The postal address of the customer.

customer_id: Option<String>

The id of the customer.