Struct google_reseller1_sandbox::CustomerMethods

pub struct CustomerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on customer resources. It is not used directly, but through the Reseller hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_reseller1_sandbox as reseller1_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use reseller1_sandbox::Reseller;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Reseller::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.customers();

Methods

impl<'a, C, A> CustomerMethods<'a, C, A>

fn update(&self, request: Customer, customer_id: &str) -> CustomerUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update a customer resource if one it exists and is owned by the reseller.

Arguments

request - No description provided.

customerId - Id of the Customer

fn insert(&self, request: Customer) -> CustomerInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a customer resource if one does not already exist.

Arguments

request - No description provided.

fn patch(&self, request: Customer, customer_id: &str) -> CustomerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update a customer resource if one it exists and is owned by the reseller. This method supports patch semantics.

Arguments

request - No description provided.

customerId - Id of the Customer

fn get(&self, customer_id: &str) -> CustomerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a customer resource if one exists and is owned by the reseller.

Arguments

customerId - Id of the Customer