Struct google_reseller1_sandbox::Seats

pub struct Seats {
    pub kind: Option<String>,
    pub number_of_seats: Option<i32>,
    pub maximum_number_of_seats: Option<i32>,
    pub licensed_number_of_seats: Option<i32>,
}

JSON template for subscription seats.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

change seats subscriptions (request)

Fields

kind

Identifies the resource as a subscription change plan request.

number_of_seats

Number of seats to purchase. This is applicable only for a commitment plan.

maximum_number_of_seats

Maximum number of seats that can be purchased. This needs to be provided only for a non-commitment plan. For a commitment plan it is decided by the contract.

licensed_number_of_seats

Read-only field containing the current number of licensed seats for FLEXIBLE Google-Apps subscriptions and secondary subscriptions such as Google-Vault and Drive-storage.