Struct google-reseller1_sandbox::Subscription

pub struct Subscription {
    pub renewal_settings: Option<RenewalSettings>,
    pub sku_id: Option<String>,
    pub kind: Option<String>,
    pub trial_settings: Option<SubscriptionTrialSettings>,
    pub transfer_info: Option<SubscriptionTransferInfo>,
    pub resource_ui_url: Option<String>,
    pub purchase_order_id: Option<String>,
    pub creation_time: Option<String>,
    pub status: Option<String>,
    pub plan: Option<SubscriptionPlan>,
    pub seats: Option<Seats>,
    pub subscription_id: Option<String>,
    pub billing_method: Option<String>,
    pub customer_id: Option<String>,
}

JSON template for a subscription.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert subscriptions (request|response)

delete subscriptions (none)

get subscriptions (response)

list subscriptions (none)

start paid service subscriptions (response)

change seats subscriptions (response)

suspend subscriptions (response)

activate subscriptions (response)

change renewal settings subscriptions (response)

change plan subscriptions (response)

Fields

renewal_settings

Renewal settings of the subscription.

sku_id

Name of the sku for which this subscription is purchased.

kind

Identifies the resource as a Subscription.

trial_settings

Trial Settings of the subscription.

transfer_info

Transfer related information for the subscription.

resource_ui_url

Ui url for subscription resource.

purchase_order_id

Purchase order id for your order tracking purposes.

creation_time

Creation time of this subscription in milliseconds since Unix epoch.

status

Status of the subscription.

plan

Plan details of the subscription

seats

Number/Limit of seats in the new plan.

subscription_id

The id of the subscription.

billing_method

Billing method of this subscription.

customer_id

The id of the customer to whom the subscription belongs.