Struct google_reseller1_sandbox::SubscriptionMethods

pub struct SubscriptionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on subscription resources. It is not used directly, but through the Reseller hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_reseller1_sandbox as reseller1_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use reseller1_sandbox::Reseller;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Reseller::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `activate(...)`, `change_plan(...)`, `change_renewal_settings(...)`, `change_seats(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `start_paid_service(...)` and `suspend(...)`
// to build up your call.
let rb = hub.subscriptions();

Methods

impl<'a, C, A> SubscriptionMethods<'a, C, A>

fn insert(&self, request: Subscription, customer_id: &str) -> SubscriptionInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates/Transfers a subscription for the customer.

Arguments

request - No description provided.

customerId - Id of the Customer

fn delete(&self, customer_id: &str, subscription_id: &str, deletion_type: &str) -> SubscriptionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels/Downgrades a subscription.

Arguments

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

deletionType - Whether the subscription is to be fully cancelled or downgraded

fn get(&self, customer_id: &str, subscription_id: &str) -> SubscriptionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a subscription of the customer.

Arguments

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn list(&self) -> SubscriptionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists subscriptions of a reseller, optionally filtered by a customer name prefix.

fn change_renewal_settings(&self, request: RenewalSettings, customer_id: &str, subscription_id: &str) -> SubscriptionChangeRenewalSettingCall<'a, C, A>

Create a builder to help you perform the following task:

Changes the renewal settings of a subscription

Arguments

request - No description provided.

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn start_paid_service(&self, customer_id: &str, subscription_id: &str) -> SubscriptionStartPaidServiceCall<'a, C, A>

Create a builder to help you perform the following task:

Starts paid service of a trial subscription

Arguments

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn change_seats(&self, request: Seats, customer_id: &str, subscription_id: &str) -> SubscriptionChangeSeatCall<'a, C, A>

Create a builder to help you perform the following task:

Changes the seats configuration of a subscription

Arguments

request - No description provided.

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn suspend(&self, customer_id: &str, subscription_id: &str) -> SubscriptionSuspendCall<'a, C, A>

Create a builder to help you perform the following task:

Suspends an active subscription

Arguments

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn activate(&self, customer_id: &str, subscription_id: &str) -> SubscriptionActivateCall<'a, C, A>

Create a builder to help you perform the following task:

Activates a subscription previously suspended by the reseller

Arguments

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer

fn change_plan(&self, request: ChangePlanRequest, customer_id: &str, subscription_id: &str) -> SubscriptionChangePlanCall<'a, C, A>

Create a builder to help you perform the following task:

Changes the plan of a subscription

Arguments

request - No description provided.

customerId - Id of the Customer

subscriptionId - Id of the subscription, which is unique for a customer