Struct google-reseller1_sandbox::SubscriptionMethodsBuilder

pub struct SubscriptionMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on subscription resources. It is not used directly, but through the Reseller hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-reseller1_sandbox" as reseller1_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use reseller1_sandbox::Reseller;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Reseller::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `activate(...)`, `change_plan(...)`, `change_renewal_settings(...)`, `change_seats(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `start_paid_service(...)` and `suspend(...)`
// to build up your call.
let rb = hub.subscriptions();

Methods

impl<'a, C, NC, A> SubscriptionMethodsBuilder<'a, C, NC, A>

fn insert(&self, request: &Subscription, customer_id: &str) -> SubscriptionInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates/Transfers a subscription for the customer.

fn delete(&self, customer_id: &str, subscription_id: &str, deletion_type: &str) -> SubscriptionDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Cancels/Downgrades a subscription.

fn get(&self, customer_id: &str, subscription_id: &str) -> SubscriptionGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a subscription of the customer.

fn list(&self) -> SubscriptionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists subscriptions of a reseller, optionally filtered by a customer name prefix.

fn change_renewal_settings(&self, request: &RenewalSettings, customer_id: &str, subscription_id: &str) -> SubscriptionChangeRenewalSettingMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Changes the renewal settings of a subscription

fn start_paid_service(&self, customer_id: &str, subscription_id: &str) -> SubscriptionStartPaidServiceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Starts paid service of a trial subscription

fn change_seats(&self, request: &Seats, customer_id: &str, subscription_id: &str) -> SubscriptionChangeSeatMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Changes the seats configuration of a subscription

fn suspend(&self, customer_id: &str, subscription_id: &str) -> SubscriptionSuspendMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Suspends an active subscription

fn activate(&self, customer_id: &str, subscription_id: &str) -> SubscriptionActivateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Activates a subscription previously suspended by the reseller

fn change_plan(&self, request: &ChangePlanRequest, customer_id: &str, subscription_id: &str) -> SubscriptionChangePlanMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Changes the plan of a subscription