Struct google_reseller1_sandbox::Subscription

pub struct Subscription {
    pub status: Option<String>,
    pub sku_id: Option<String>,
    pub trial_settings: Option<SubscriptionTrialSettings>,
    pub deal_code: Option<String>,
    pub resource_ui_url: Option<String>,
    pub customer_domain: Option<String>,
    pub plan: Option<SubscriptionPlan>,
    pub seats: Option<Seats>,
    pub billing_method: Option<String>,
    pub renewal_settings: Option<RenewalSettings>,
    pub transfer_info: Option<SubscriptionTransferInfo>,
    pub kind: Option<String>,
    pub suspension_reasons: Option<Vec<String>>,
    pub creation_time: Option<String>,
    pub purchase_order_id: Option<String>,
    pub subscription_id: Option<String>,
    pub customer_id: Option<String>,
}

JSON template for a subscription.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert subscriptions (request|response)

delete subscriptions (none)

get subscriptions (response)

list subscriptions (none)

start paid service subscriptions (response)

change seats subscriptions (response)

suspend subscriptions (response)

activate subscriptions (response)

change renewal settings subscriptions (response)

change plan subscriptions (response)

Fields
status: Option<String>

Status of the subscription.

sku_id: Option<String>

Name of the sku for which this subscription is purchased.

trial_settings: Option<SubscriptionTrialSettings>

Trial Settings of the subscription.

deal_code: Option<String>

External name of the deal, if this subscription was provisioned under one. Otherwise this field will be empty.

resource_ui_url: Option<String>

Ui url for subscription resource.

customer_domain: Option<String>

Primary domain name of the customer

plan: Option<SubscriptionPlan>

Plan details of the subscription

seats: Option<Seats>

Number/Limit of seats in the new plan.

billing_method: Option<String>

Billing method of this subscription.

renewal_settings: Option<RenewalSettings>

Renewal settings of the subscription.

transfer_info: Option<SubscriptionTransferInfo>

Transfer related information for the subscription.

kind: Option<String>

Identifies the resource as a Subscription.

suspension_reasons: Option<Vec<String>>

Read-only field containing an enumerable of all the current suspension reasons for a subscription. It is possible for a subscription to have many concurrent, overlapping suspension reasons. A subscription's STATUS is SUSPENDED until all pending suspensions are removed. Possible options include: - PENDING_TOS_ACCEPTANCE - The customer has not logged in and accepted the Google Apps Resold Terms of Services.
- RENEWAL_WITH_TYPE_CANCEL - The customer's commitment ended and their service was cancelled at the end of their term.
- RESELLER_INITIATED - A manual suspension invoked by a Reseller.
- TRIAL_ENDED - The customer's trial expired without a plan selected.
- OTHER - The customer is suspended for an internal Google reason (e.g. abuse or otherwise).

creation_time: Option<String>

Creation time of this subscription in milliseconds since Unix epoch.

purchase_order_id: Option<String>

Purchase order id for your order tracking purposes.

subscription_id: Option<String>

The id of the subscription.

customer_id: Option<String>

The id of the customer to whom the subscription belongs.